+++ src/zarr/registry.py
@@ -22,17 +22,20 @@ if TYPE_CHECKING:
     )
     from zarr.abc.numcodec import Numcodec
     from zarr.core.buffer import Buffer, NDBuffer
+    from zarr.core.chunk_grids import ChunkGrid
     from zarr.core.chunk_key_encodings import ChunkKeyEncoding
     from zarr.core.common import JSON
 
 __all__ = [
     "Registry",
     "get_buffer_class",
+    "get_chunk_grid_class",
     "get_chunk_key_encoding_class",
     "get_codec_class",
     "get_ndbuffer_class",
     "get_pipeline_class",
     "register_buffer",
+    "register_chunk_grid",
     "register_chunk_key_encoding",
     "register_codec",
     "register_ndbuffer",
@@ -63,6 +66,7 @@ __codec_registries: dict[str, Registry[Codec]] = defaultdict(Registry)
 __pipeline_registry: Registry[CodecPipeline] = Registry()
 __buffer_registry: Registry[Buffer] = Registry()
 __ndbuffer_registry: Registry[NDBuffer] = Registry()
+__chunk_grid_registry: Registry[ChunkGrid] = Registry()
 __chunk_key_encoding_registry: Registry[ChunkKeyEncoding] = Registry()
 
 """
@@ -103,6 +107,11 @@ def _collect_entrypoints() -> list[Registry[Any]]:
     data_type_registry._lazy_load_list.extend(entry_points.select(group="zarr.data_type"))
     data_type_registry._lazy_load_list.extend(entry_points.select(group="zarr", name="data_type"))
 
+    __chunk_grid_registry.lazy_load_list.extend(entry_points.select(group="zarr.chunk_grid"))
+    __chunk_grid_registry.lazy_load_list.extend(
+        entry_points.select(group="zarr", name="chunk_grid")
+    )
+
     __chunk_key_encoding_registry.lazy_load_list.extend(
         entry_points.select(group="zarr.chunk_key_encoding")
     )
@@ -125,6 +134,7 @@ def _collect_entrypoints() -> list[Registry[Any]]:
         __pipeline_registry,
         __buffer_registry,
         __ndbuffer_registry,
+        __chunk_grid_registry,
         __chunk_key_encoding_registry,
     ]
 
@@ -154,6 +164,10 @@ def register_ndbuffer(cls: type[NDBuffer], qualname: str | None = None) -> None:
 
 def register_buffer(cls: type[Buffer], qualname: str | None = None) -> None:
     __buffer_registry.register(cls, qualname)
+
+
+def register_chunk_grid(key: str, cls: type) -> None:
+    __chunk_grid_registry.register(cls, key)
 
 
 def register_chunk_key_encoding(key: str, cls: type) -> None:
@@ -294,6 +308,16 @@ def get_ndbuffer_class(reload_config: bool = False) -> type[NDBuffer]:
     raise BadConfigError(
         f"NDBuffer class '{path}' not found in registered buffers: {list(__ndbuffer_registry)}."
     )
+
+
+def get_chunk_grid_class(key: str) -> type[ChunkGrid]:
+    __chunk_grid_registry.lazy_load(use_entrypoint_name=True)
+    if key not in __chunk_grid_registry:
+        raise KeyError(
+            f"Chunk grid '{key}' not found in registered chunk grids: "
+            f"{list(__chunk_grid_registry)}."
+        )
+    return __chunk_grid_registry[key]
 
 
 def get_chunk_key_encoding_class(key: str) -> type[ChunkKeyEncoding]:
